ice_breakers/requirements.txt
pyspark
langchain
langchain-openai
langchain-community
langchainhub
python-dotenv